// src/types.ts
export type SiteRow = {
	id?: string;
	officialName?: string;
	nickname?: string;
	operator?: string;
	servicePopulation?: string;
	serviceCategory?: string;
	admissionType?: string;
	lockedStatus?: string;
	dropIn?: string;
	dropInNotes?: string;
	dropOff?: string;
	dropOffNotes?: string;
	acceptedFrom?: string;
	acceptedFromNotes?: string;
	intakeComplexity?: string;
	medicalClearance?: string;
	eligibilitySummary?: string;
	lengthOfStay?: string;
	bedsDph?: string;
	pitCapacity?: string;
	capacityConstraints?: string;
	programAddress?: string;
	intakeAddress?: string;
	waitLocation?: string;
	hoursOperation?: string;
	hoursIntake?: string;
	transportIn?: string;
	transportOut?: string;
	transportSupport?: string;
	serviceRegion?: string;
	neighborhood?: string;
	primaryServiceArea?: string;
	contact?: string;
	mvp?: string;
	adaAccessible?: string;
	keyAccessNotes?: string;
};